integration the utxo test and delete Coin_test.go
